Reduced duplicated code
